Add history attributes to ChargeVersionModel
https://eaflood.atlassian.net/browse/WATER-4560

> Part of the work to display NALD mod log records as history in WRLS

So, we are now importing mod logs from NALD and have added a model linked to charge, licence and return version models to allow us to retrieve them from the database.

We are in the process of building [a new licence history page](https://github.com/DEFRA/water-abstraction-system/pull/1182) and we need to update the view of the charge version in the future.

Because of this, we believe adding a modifier and instance methods to the `ChargeVersionModel` that will handle fetching the mod logs and determining this information is valuable.

This change updates the `ChargeVersionModel` with these properties. This is similar to the work we did in [Add history attributes to ReturnVersionModel](https://github.com/DEFRA/water-abstraction-system/pull/1267) and [Add history attributes to LicenceVersionModel](https://github.com/DEFRA/water-abstraction-system/pull/1271).