Ignore new `/import/licence` endpoint in logs
https://eaflood.atlassian.net/browse/WATER-4575

> Part of the work to replace the legacy licence import in readiness for ReSP

In [Import Licence](https://github.com/DEFRA/water-abstraction-system/pull/1191), we added our first steps of migrating the legacy licence import from [water-abstraction-import](https://github.com/DEFRA/water-abstraction.import).

At the same time we added some debug logging only to realise later [they were blowing up the logs](https://github.com/DEFRA/water-abstraction-system/pull/1266)!

That's all sorted now, but we are still seeing a lot of traffic nightly appearing in the logs. It is to be expected because  [water-abstraction-import](https://github.com/DEFRA/water-abstraction-import) is pinging it for every licence it is importing (which is almost 80K! 😱)

We will be doing some form of logging when the [next iteration of the import](https://github.com/DEFRA/water-abstraction-system/pull/1195) is approved and merged.

But till then we don't need to see that the endpoint was hit 80K times each night.

This change adds the route to our `HapiPinoIgnoreRequestService` which is used to control which routes are ignored in the logs.