Reverting presenter test with tweaks
I think because of the churn that has gone on with mod logs whilst trying to pull together this PR, changes have been made to the presenter tests that are not really needed.

Plus, we can create model instances from an Object, avoiding the need to create 'fake' versions of methods. So, we revert this test file pretty much back to its original state, before then fixing and expanding the tests.