Alan's Update view return version with mod log (#1276)
Alan's suggested updates

* Reverting presenter test with tweaks

I think because of the churn that has gone on with mod logs whilst trying to pull together this PR, changes have been made to the presenter tests that are not really needed.

Plus, we can create model instances from an Object, avoiding the need to create 'fake' versions of methods. So, we revert this test file pretty much back to its original state, before then fixing and expanding the tests.

* Housekeeping - remove redundant describe

Made me keep putting the `only()` in the wrong place.

* Fix the tests I broke switching to a model

I'd overlooked switching to using an actual model instance rather than just an object would mean ensuring we have model instances all the way down!

* Properly handle all reason scenarios

The notes explain what the scenarios are we need to cover.

* Expand reason tests

We are not looking to retest what is already covered by the model. But we do want to test every scenario our presenter may face. Plus this serves as a handy way to document all the possible outcomes we may get from the presenter.

* Move createdBy to a function

IMHO I think this cleans up what is happening in the go(). It also avoids doing the same 'work' twice. `$createdBy` gives us a result which it is then on the presenter to determine what to do with it.

* Remove new date wrapper

I don't think we need this any more. Did some testing and all seemed fine.

* Remove escape

Calling a macro appears to be different to a function or filter. Nunjucks interprets the value differently so we don't need to pass it to the `escape` filter.

* Revert change to seeder

It was fine as a one liner!

* Tweak to the helper function

We can revert the beforeEach back to what it was now we have a better idea of how the helper function can work.

* Revert changes to view service test then tweak

We can apply the same tweaks we did to view presenter test to get this passing again with the minimum of changes.