Fix JSDOC's lint description and type issues
Previous work https://github.com/DEFRA/water-abstraction-system/pull/1269

Building on from previous work of using eslint to handle out code conventions this change will use the auto fix feature of eslint and fix any errors relating to a jsoc description not having a - to separate and also fix any types that are incorrectly named i.e Object shoudl dbe object.

Some of the more complex changes such as adding returns and parameters to the jsdocs will be left to be done as when the file is touched or when time permits addressing tech debt.