Tell JSDoc linter to ignore private functions (#1280)
https://github.com/DEFRA/water-abstraction-team/issues/125

We recently added a new [eslint-plugin-jsdoc](https://github.com/gajus/eslint-plugin-jsdoc) plugin in [Lint JSDOC's](https://github.com/DEFRA/water-abstraction-system/pull/1269) to take our work of helping new and existing devs follow our JSDoc conventions to the next level.

For reference, we aim to document all our public functions. Forcing devs to write about why something exists is a handy way to try and slow them down and think about what they are doing!

But we don't ask anyone to document private functions, for example, `function _determineStartDate()` — basically, anything with a leading underscore. However, there are times when a comment is handy. We have to do some gnarly stuff on this service, and sometimes, we can't make the code self-documenting. 😢

At the moment, the plugin is penalising us for doing this. These comments don't include params and return statements because that was not the intent of the comment.

So, this change is about getting the plugin to ignore private methods.