Alans suggested import licence version changes
https://eaflood.atlassian.net/browse/WATER-4575

In this instance, we felt it was more constructive to provide the feedback for the original [Import Licence versions](https://github.com/DEFRA/water-abstraction-system/pull/1195) proposed PR as a series of commits from where the code has been left.

This can be used to guide suggested changes, hopefully with better context than just PR comments.