Make route specific to legacy import licence
The plan is to add a separate route for the ReSP version when we are ready to import licences using that way. In the meantime, there is no reason to make it clear that this endpoint is to support importing a licence using NALD data. It also explains the call to `ImportLegacyProcessLicenceService` in the controller.

> This will mean tweaking water-abstraction-import again but we think it is worth doing so now rather than later.