Update query in fetch licence service
The thinking behind this is that the database server is the one with by far the most 'oomph' in our environment. That resource plus the power of SQL means we can get it to do some of the work and thus avoid having to do it on our meagre EC2 instance!

That comes through in this change in two key areas.

- where a value is 'null' have the query return us a _true_ NULL, thus avoiding checks for `foo === 'null'` in our code
- where a value is another type, have PostgreSQL do the conversion, for example, Dates

Because we're using `db.raw` (which we have to because Knex snake case mappers baulk at the NALD column names!) we have to live with snake_case properties. But there is nothing stopping us from renaming the fields in the queries to something more meaningful. In this way, the terms 'AREP_AREA_CODE' or 'FGAC_REGION_CODE' can be consigned to a single place in our code.

The final tweak is the sub query to identify the what the earliest start licence version start date. This one allows us to remove the dependence on the licence versions result and helps simplify our presenter logic. Plus again, the DB has more oomph to work that out!