Update Licence Presenter following fetch changes
With the changes we made to the query in FetchLicenceService, we can remove all the type casting the presenter was doing, along with changing all the property references.

We can also simplify how we determine the start date.

Beyond that we just do some housekeeping to the comments and the file.