Update licence validator
We've not updated how the validation is done. We've just simplified what happens when an error happens. The Joi error has some juicy detail that might be helpful if we have to debug an issue. So, rather than wrap the error message in a new Error we pass the original Joi ValidationError back.

We also move schema into the function. Instantiating variables outside the scope of a function (constants being the exception) is not a convention we follow. It could have been promoted to a constant. But there is no reason for it not to follow our other validations and move it into the `go()` where it is used.

We finish with some housekeeping in the file.