Update Fetch licence vers. + split out purposes
We make the same changes to the query to get the DB to do the work of cleaning up the data and casting to proper data types as we did in `FetchChargeVersions`.

But the key change is we are taking out fetching the licence version purposes from the query. If performance was the critical aspect, we'd be pushing this query into `FetchLicenceService`. One request to the DB would fetch all the data we need for the licence in one go. But this legacy version of the job will be run in the background over night, so we can be relaxed about how long it takes per licence.

When this goes to ReSP, we will be forced to extract the parts of the licence in separate API calls. As well as consistency, the final argument to make for splitting is it allows us to simplify both the presenter and validator that come later. Each now is responsible for a single part of the process, making it easier to understand, test and maintain going forward.