Update Licence vers. presenter to handle 1 record
Now the licence versions no longer contain the purposes within them we need to make changes to the presenter. But along with splitting out the entity types, we're also going to move the functionality to look at single records rather than arrays of them.

Again it comes back to making each thing (fetch, presenter, validator) that little bit less complex and easier to test.

This also helps shift the import process to be one of transformation, broken into steps. Once we've made the changes to the each of those module type, and added the ones for licence version purposes, we'll bring them back together in 'transformers' (mappers in disguise!)