Autobots - show yourselves!
So, having looked at what the `ProcessLicenceService` was doing, it was clearly doing things in 3. Fetching, presenting and validating. But doing that for licences, licence versions, and licence version purposes.

Because of a dependence between them, this was all being done in the process service, and we know we have to add yet more entity types to it.

The main idea behind this suggested change, is to get to a place where each entity type can be 'transformed' separately from the other. The `ProcessLicenceService` then is purely an orchestrator of the transformation steps and as we add more entities, we are just adding another transformer for the service to call.

Once all parts have been transformed, we can bring them back together to check the structure as a whole, before persisting within one DB transaction. Well, that's the aim anyway (kind of coding as I go!)