Start moving work out of persist services
Currently, the persist services do the work of identifying reference records, for example, region.

It is doing this using data we're importing in the fetch services. So, by altering the fetch service queries we can have it 'fetch' this information at the same time.

We can even then use the validators to confirm we have a valid UUID.