Build the licence structure
We're now using the transformers to 'build' the licence structure as we go.

With our transformed licence, we can pass it to the licence version transformer, and that will attach the versions to the licence.

We can then pass our transformed licence again, now into the licence version purposes transformer. Not only will it transform and validate the data from NALD, it will match the licence version purpose to its version (we tweak the query to make that a little cleaner).

TBH, I was originally intending to do this as a stage _after_ transformation. But we don't pull back issue and increment values when fetching the licence version purpose, which are needed for matching to the licence version.

I could have done (have the presenter add them to the transformed object), but then somewhere later in the process they would need to be stripped off the object else an error would happen when you try and persist the record.

This felt cleaner. Plus, I like the idea of the licence being 'built up' as we walk through each transformation step.