Update licence validator tests
Because of changes we've made to the data being fetched we obviously had to make changes to the validator. But 2 other key changes to note.

We spotted that passing in a date property with a value of 1 was valid because Joi is automatically converting the value to a date. So, we switch of automatic conversion to ensure Joi doesn't mess with our values in an attempt to be helpful.

The second is moving changes to our test data out of the `it()` and into a `beforeEach()`. This is a convention/pattern we've highlighted before: Arrange, Act and Assert.

Using a `beforeEach()` makes the Arrange stage explicit and the Act stage consistent throughout the tests.