Remove the fixtures
Perhaps when each of the services was doing more, and therefore each test file needed more data the fixtures really helped. But on reflection may be that should have been seen as a sign that they needed simplifying.

With the tests split out we've been able to use the exist pattern of generating a subset of data using a function in the test file. `PersistLicenceService` is probably the one service left that would have benefited. But it seems overkill to keep them around for one test file.

I may regret this! But for now, I think we can park adding fixtures.