Add validation to reject whitespace in licence ref
I can point to numerous bug tickets which have been caused by the issue described in the note. To the extent we now have a hidden endpoint which allows us to quietly purge these problem records from the service without needing to wait for a new release and fix migration.

We also don't want to automatically trim the whitespace and carry on. It could be an error and either we were never meant to see the licence record, or the it is for a different licence (and we'd end up updating the wrong one).