Flag Supplementary 2PT for editing returns
https://eaflood.atlassian.net/browse/WATER-4199

As part of the ongoing work for the two-part tariff supplementary bill runs, we need to be able to flag the licence years when a return has been edited or submitted. The UI deals with the initial request to system so this PR is focusing on taking in a returnID and working out the years that need to be flagged on the licence.