Fix registered to link in view licence
https://eaflood.atlassian.net/browse/WATER-4650

The new view licence page went live in the [release today](https://github.com/DEFRA/water-abstraction-system/releases/tag/v0.20.0). But users have found if a licence has a registered user, and you click the link to see their status, it results in a 'page not found'.

Having looked at the issue, it appears we have lost adding the user ID to the link.

This change fixes the issue.