Hmm ...
Having dug into the `registeredToAndLicenceName()` modifier I can see it is pulling out the correct name to display.

But the view is expecting something to set `primaryUser.userId`. But double checking the existing code and the original PR there is no other reference made to this.

In fact, to pull out the user ID for the registered to user, you need to be able to link `public.licence_entities` to `public.users`. However, this is impossible because we are not including `idm.users.external_id` in our `public.users` view!

So, this isn't a bug that's crept in, this has been here since day 1!

Doh! 😱🤦