Replace registeredToAndLicenceName
With us having to go even deeper to retrieve the primary user ID in order to make the broken link work we've decided to split up the existing modifier/helper.

Having them separate means it is easier to see what is needed to determine each property. Having done so it becomes clear licence name is a doddle but primary user is a hell scape!