Fix up FetchLicenceService
We remove the data transformation from the service so we can leave that to the presenter and simplify this fetch service.

In the tests, the modifiers are already covered by tests in the `LicenceModel` itself so we don't need to re-test them here.

We just add enough, i.e. a workflow record, to cover those query elements we're doing in the fetch service itself.

We also tidy up field names, JSDoc comments and any sorting.