Housekeeping - add basic access user to seed
Needed to confirm that the basic access internal user has no roles assigned to it. But to do so we needed to create a new user account because we are not currently seeding one.

So, updated the seed whilst about it to ensure we have the ability to access a user with this permission level both in tests and our non-prod environments.