Fix up view licence presenter
Found we were returning values that either were not used by the view, or that could be simplified.

In the tests, we were generating data that the presenter ignored or had tests that covered functionality already covered by other tests.

So, in all we were able to simplify both the presenter and the tests. On top of that we brought the naming of the functions et al in line with our conventions plus some other housekeeping.