Override "Missing JSDoc comment" rule for controllers (#1292)
https://github.com/DEFRA/water-abstraction-team/issues/125

We use skinny controllers that don't do much more than call a service and return a view. We have therefore decided that it is not necessary to add JSDoc comments to our controllers.

This PR will add an override to the linting rules so that JSDoc comments are not required for the controllers.