Add select createdAt to Return Version
An interesting thing to note here is the need to all the other values we depend on from the parent record.

When there was no `select()` it was doing the equivalent of `select(*)`. So, you could say there was no need for this change because `createdBy` was always being returned anyway.

But this makes the intent clearer and hopefully will mean no one gets an unintended surprise because a query they are running that incorporates these modifiers includes unexpected fields.