Add createdAt to Version history modifiers (#1293)
https://eaflood.atlassian.net/browse/WATER-4566

> Part of the work to display NALD mod log records as history in WRLS

We recently added a new `history()` modifier to the charge, licence, and return version models that allows us to pull through the mod log and, in some cases, user details when querying for a record. We then call `$` instance methods to determine who created a record or what the reason was.

We're using it in [Build new licence history page](https://github.com/DEFRA/water-abstraction-system/pull/1182), but that change has highlighted that to make `$createdAt()` work, we have to ensure we include `createdAt` in the parent query's `select()`.

This should be dealt with in the `history()` modifier. This change makes it so!