Build new licence history page (#1182)
* Build new licence history page

https://eaflood.atlassian.net/browse/WATER-4566

Billing & Data rely on the NALD ‘mod log’ to be able to see that all 3 legs are in sync when managing a licence.
For example, when a licence is varied (a new licence version is created), new charge and return versions are also created (if applicable).

This PR is focused on building a new 'History' page to replicate ‘mod log’ within our system. It should display all licence,
charge and return versions for the licence in a single table (no pagination), ordered by created date in descending order.