Fix return version tests
As part of ongoing work to remove database cleans from the test suit we are getting intermittent tests failing mainly due to unique constraints where the range is too small to ensure uniqueness.

This change adds a generate random region code function explaining the issue.

The aim of this change is to encapsulate the issue and remove any need to explain the issue else where in the tests.