Fix user model tests (#1296)
As part of ongoing work to remove database cleans from the test suit we are getting intermittent tests failing. In this case we are seeing tests interfering with each other when they use the user id's form the ref data.

As these are fixed (seeded) there maybe some cross over in tests. So we update the tests to find the individual elements it has created and asserts against that, rather than length of an array or position of the element in the array.