Fix charge category charge ref tests
As part of ongoing work to remove database cleans from the test suit we are getting intermittent tests failing. In this case we are seeing tests interfering with each other when they use the generated charge reference in the ChargeCategoryHelper.

A charge version is made up of three parts. Some with a small range.

This change will broaden the range of the charge version but attempt to document the parts of the charge reference and why there is a defined range for the parts.