Fix licence workflow warning on view licence page
https://eaflood.atlassian.net/browse/WATER-4652

In [Add licence workflow warning to view licence page](https://github.com/DEFRA/water-abstraction-system/pull/1204) we added a new warning message to the view licence page to indicate that a change has been made to a licence and so it has been added to workflow.

But our users have spotted it's show for licences that _don't_ have anything in workflow! 😱

We've tracked the problem down to the fact workflow uses a mix of hard and soft deleted records. When we are checking if a licence is in there, we are not discounting soft deleted records.

This change fixes the query we use to resolve this.