Fix licence workflow warning on view licence page (#1300)
https://eaflood.atlassian.net/browse/WATER-4652

In [Add licence workflow warning to view licence page](https://github.com/DEFRA/water-abstraction-system/pull/1204), we added a new warning message to the view licence page to indicate that a licence change has been made and that it has been added to the workflow.

But our users have spotted it's showing for licences that _don't_ have anything in workflow! 😱

We've tracked the problem down to the workflow's use of a mix of hard and soft deleted records. When we check if a licence is in there, we do not discount soft deleted records.

This change fixes the query we used to resolve this.