Add relation mapping to `LicenceModel` https://eaflood.atlassian.net/browse/WATER-4587
In order to get the data required to display the two-part tariff supplementary banner the query in `app/services/licences/fetch-licence.service.js` is going to be updated to include data from the `licence_supplementary_years` table.

However, whilst there is currently a relation mapping to the 'LicenceModel' from the `LicenceSupplementaryYearModel`. There is no relation mapping to the `LicenceSupplementaryYearModel` from the `LicenceModel` so it will be added in this PR.