Update Banners to incorporate 2PT Supplementary https://eaflood.atlassian.net/browse/WATER-4587
Now that we are introducing the new bill run type, we will need to update the banner options shown on the other licence pages, 2PT Supplementary Billing will have it’s own banner which could be combined with the 2 existing banners at any point.

In this PR the new banners are going to be created and added to the view licence page.