Add external id to companies view.
https://eaflood.atlassian.net/browse/WATER-4659

As part of the ongoing to work to migrate the import service into the water-abstraction-system, we need to add the external id for the companies (know as parties in NALD).

When we import the companies (parties) we need to upsert the data. The current import service uses the external id to resolve the conflict.

This follows the same approach as the other import work.