Trigger 2PT Supplementary Bill Run - Pt2 (#1262)
https://eaflood.atlassian.net/browse/WATER-4631

As we treat 2PT supplementary as a separate bill run we will need a new journey to trigger it.

The first part of this work was carried out in PR https://github.com/DEFRA/water-abstraction-system/pull/1259

In this PR the remainder of the journey to enable the triggering of a 2PT supplementary will be completed. This will include selecting the required financial year and triggering the bill run.