Fix incorrect edit link for void returns
https://eaflood.atlassian.net/browse/WATER-4661

Our users have spotted that we are providing a link for `VOID` returns in the view licence page that takes users to the returns edit page. Users should not be able to edit a `VOID` return and unfortunately there is nothing in the legacy edit page to stop this.

So, we need to fix the link to take users to view returns page, not edit.