Ensure we have proper coverage of the presenter
Before we start making changes to the presenter we want to ensure we have proper coverage of everything it is doing now.

One thing it has highlighted is that we are marking overdue a return that is due 'today'. This is because when we generate the variable `today` in the presenter we are not discarding the time.

So, for example, if the return is due 2024-09-05 and the date is the same, when we do the comparison we are actually doing `2024-09-05 00:00:00 < 2024-09-05 13:27:56`.

This means it display as "overdue" when in fact it should be "due" (confirmed by comparing to the legacy version of the returns tab).