Fix incorrect edit link for void returns (#1311)
https://eaflood.atlassian.net/browse/WATER-4661

Our users have noticed that we provide a link for `VOID` returns on the view licence page that takes them to the returns edit page. Users should not be able to edit a `VOID` return; unfortunately, there is nothing on the legacy edit page to stop this.

So, we need to fix the link to take users to view returns page, not edit.

---

While working on this, we also found that the status was showing return logs due 'today' incorrectly (they were showing as "overdue" when the status should be "due").