Fix import licence log silencer (#1313)
https://eaflood.atlassian.net/browse/WATER-4575

> Part of the work to replace the legacy licence import in readiness for ReSP

Argh! We've done it again! We made a change to [Ignore new /import/licence endpoint in logs](https://github.com/DEFRA/water-abstraction-system/pull/1275), so the logs became silent again instead of outputting 80K requests!

But then in [Import Licence versions](https://github.com/DEFRA/water-abstraction-system/pull/1195) we tweaked the route, and guess what we forgot to do! 😱🤦

This change updates the `HapiPinoIgnoreRequest` with the updated route.