Update Return req set up journey to use new points (#1301)
https://eaflood.atlassian.net/browse/WATER-4600
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

We've been extending and amending the import of return versions from NALD to WRLS as part of our work to switch from NALD to WRLS to manage them. To support this, we [Created a return-requirement-points table](https://github.com/DEFRA/water-abstraction-service/pull/2540) and [updated the import](https://github.com/DEFRA/water-abstraction-import/pull/933) to import them point details.

Users select these points as part of the return requirements setup journey we've built. We currently extract them from the JSON blob stored in the `permit.licence` table. However, the import service only populates the points for licences.

- That have not ended
- That have a current licence version

Otherwise, `permit.licence` is not populated with the points data our journey relies on, causing it to throw an error.

So, we needed another source for points. In [Start importing licence version points to 'water'](https://github.com/DEFRA/water-abstraction-import/pull/1009) we updated the **water-abstraction-import** to start importing licence version purpose points to a new table.

We added the model to access the data in [Add Licence Version Purpose Point model](https://github.com/DEFRA/water-abstraction-system/pull/1288). In [Add base point model with a describe helper](https://github.com/DEFRA/water-abstraction-system/pull/1289), we extended the models to show how to describe a point.

With all this in place, we can now update the return requirements setup journey to source 'points' from the new table. This will fix the issue we currently have with it and get it working again for all licences.

The fix requires

- updating the 'use abstraction data' journey to pull points from the new table
- updating the 'manual journey' to pull points from the new table

We'll also update the 'copy from existing' journey to use the new features of the `BasePointModel` rather than the general lib helper.