Fix - Use NALD point ID instead of ID
Working on [Update rtn requirement fixtures to use new points](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/134) exposed that copy from an existing return requirement was not actually working.

We tracked the issue down to the use of IDs. The changes we had made to the journey to support reading points from `water.licence_version_purpose_points` were the cause.

We were storing the licence point ID in the session and using it as the value when displaying the points on the manual journey. But when you copy from existing, the source point comes from `water.return_requirement_points`. When it got to the `/check` page its presenter was trying to match the ID we'd stored in the session (from return points) with the ID of a licence point. No matching was being found causing the error.

Fortunately, the one value both tables share is the NALD point ID. Like with our journey, when you create the return requirement in NALD you select from the same licence points as we do. This means there will _always_ be a match between licence points and return points. You just have to match on NALD Point ID instead.