Fix use of new points in setup ret. req. journey (#1314)
https://eaflood.atlassian.net/browse/WATER-4600
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

We've been extending and amending the import of return versions from NALD to WRLS as part of our work to switch from NALD to WRLS to manage them.

In [Update Return req set up journey to use new points](https://github.com/DEFRA/water-abstraction-system/pull/1301) we updated the return requirement setup journey to use the new `water.licence_version_purpose_points` table we've created and now populate thanks to [a change to the import](https://github.com/DEFRA/water-abstraction-system/pull/1301).

We then began updating our acceptance tests [to update the return requirement fixtures to use new points](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/134) but found the 'Copy from existing journey' was failing.

We tracked the issue down to the use of IDs. The changes we had made to the journey to support reading points from `water.licence_version_purpose_points` were the cause.

Starting with the manual journey, we updated the fetch points service to retrieve points information from `water.licence_version_purpose_points`. The ID we provided in the presenter to the view was the `LicenceVersionPurposePoint.id`. This is then what gets stored in the session when a user makes a selection.

Then, we did something similar in the 'Use abstraction data' journey. Again, we retrieve the licence version purpose points and store their IDs in the session.

In both journeys, when you get to the `/check` screen, it will use the `FetchPointsService` to fetch the licence version purpose points again. The presenter for the check page can match the selected points in the session to the results of the fetch points service.

The journey we broke was 'Copy from existing'. In that scenario, the points are sourced from `water.return_requirement_points`, which means they have a different ID than what is in `water.licence_version_purpose_points` that we are storing in the session. This means that when you then get passed to the `/check` page, it is unable to find a match in the results from `FetchPointsService,` so an error is thrown.

Fortunately, the one value both tables share is the NALD point ID. Like with our journey, when you create the return requirement in NALD, you select from the same licence points as ours, which also have a NALD point ID. This means there will _always_ be a match between licence points and return points. You just have to match on NALD Point ID instead of their IDs.